ENG-770: Subject is a datastructure, and open_subjects is not an author door
models.py declares mapped rows; Subject is a value type, so it moves to
druks/durable/datastructures.py beside the other frozen ones. Its id comes
first.

open_subjects stays on the internal read side. Nothing calls it yet, and the
author surface is not where an unused read belongs.
